3D_PETRAUL/requirements.txt
streamlit
numpy
matplotlib
pandas
Pillow
plotly
uuid